Split out dirty/save timer logic, added save-ability on note edit
