+++ app/actions/editor.tsx
@@ -1,5 +1,8 @@
+import Redux from 'redux'
 import {SetDirtyAction, SetDirtyTimeoutAction, SetLineAction} from './ActionTypes'
 import {PanelType} from '../reducers/StateTypes'
+import {store} from '../store'
+import {saveQuest} from './quest'
 
 export function setDirty(is_dirty: boolean): SetDirtyAction {
   return {type: 'SET_DIRTY', is_dirty};
@@ -19,4 +22,32 @@ export function setOpInit(mathjs: string) {
 
 export function panelToggle(panel: PanelType) {
   return {type: 'PANEL_TOGGLE', panel};
+}
+
+export function updateDirtyState(): ((dispatch: Redux.Dispatch<any>)=>any) {
+  return (dispatch: Redux.Dispatch<any>): any => {
+    const editor = store.getState();
+    if (!editor.dirty) {
+      dispatch(setDirty(true));
+    }
+
+    if (editor.dirtyTimeout) {
+      clearTimeout(editor.dirtyTimeout);
+    }
+
+    const timer = setTimeout(function() {
+      const state = store.getState();
+      // Check if a future state update overrode this timer.
+      if (state.editor.dirtyTimeout !== timer) {
+        return;
+      }
+      // Check the store directly to see if we're still in a dirty state.
+      // The user could have saved manually before the timeout has elapsed.
+      if (state.editor.dirty) {
+        dispatch(saveQuest(state.quest));
+      }
+      dispatch(setDirtyTimeout(null));
+    }, 2000);
+    dispatch(setDirtyTimeout(timer));
+  }
 }
