+++ app/components/NotesPanelContainer.tsx
@@ -2,7 +2,7 @@ import Redux from 'redux'
 import {connect} from 'react-redux'
 import {AppState} from '../reducers/StateTypes'
 import NotesPanel, {NotesPanelDispatchProps, NotesPanelStateProps} from './NotesPanel'
-import {setOpInit} from '../actions/editor'
+import {updateDirtyState} from '../actions/editor'
 
 const mapStateToProps = (state: AppState, ownProps: any): NotesPanelStateProps => {
   return {
@@ -12,7 +12,10 @@ const mapStateToProps = (state: AppState, ownProps: any): NotesPanelStateProps =
 
 const mapDispatchToProps = (dispatch: Redux.Dispatch<any>, ownProps: any): NotesPanelDispatchProps => {
   return {
-    //onNotesBlur
+    onDirty: (realtime: any, text: string) => {
+      realtime.setText(text);
+      dispatch(updateDirtyState());
+    },
   };
 }
