+++ app/components/QuestIDE.tsx
@@ -3,20 +3,18 @@ import * as React from 'react'
 import {Tab} from 'material-ui/Tabs'
 
 import TextView from './base/TextView'
-import {QuestType, EditorState, AnnotationType} from '../reducers/StateTypes'
+import {AnnotationType} from '../reducers/StateTypes'
 import AppContainer from './AppContainer'
 
 export interface QuestIDEStateProps {
   realtime: any;
-  quest: QuestType;
-  editor: EditorState;
   annotations: AnnotationType[];
   lastSplitPaneDragMillis: number;
 };
 
 export interface QuestIDEDispatchProps {
-  onDirty: (realtime: any, quest: QuestType, editor: EditorState, text: string) => void;
-  onLine: (line: number, editor: EditorState) => void;
+  onDirty: (realtime: any, text: string) => void;
+  onLine: (line: number) => void;
 }
 
 interface QuestIDEProps extends QuestIDEStateProps, QuestIDEDispatchProps {}
@@ -30,8 +28,8 @@ const QuestIDE = (props: QuestIDEProps): JSX.Element => {
           realtime={props.realtime}
           annotations={props.annotations}
           lastSizeChangeMillis={props.lastSplitPaneDragMillis}
-          onChange={(text: string) => props.onDirty(props.realtime, props.quest, props.editor, text)}
-          onLine={(line: number) => props.onLine(line, props.editor)} />
+          onChange={(text: string) => props.onDirty(props.realtime, text)}
+          onLine={(line: number) => props.onLine(line)} />
       </div>
       <div className="preview">
         <AppContainer/>
